docker/requirements.txt
jsonrpclib_pelix==0.3.1
kombu==4.1.0
Django==1.11.2
celery==4.1.0
django-extensions==1.9.7
psycopg2==2.7.3.2
django-registration==2.3
django-db
django-celery-results
requests
